// app/utils/types/analytics.ts
export interface StudentAnalytics {
  total_class_leaders: number;
  total_leaders_active: number;
  total_leaders_inactive: number;
  total_leaders_approved: number;
  total_leaders_rejected: number;
}
export interface ClassroomAnalytics {
  total_classrooms: number;
  total_classroom_approved: number;
  total_classroom_pending: number;
  total_classroom_rejected: number;
}
export interface FeedbackAnalytics {
  total_reviews: number;
}